+++ react/components/EmptyDocs.tsx
@@ -1,28 +1,19 @@
 import React, { FC } from 'react'
-import {
-  FormattedMessage,
-  injectIntl,
-  defineMessages,
-  InjectedIntlProps,
-} from 'react-intl'
-
+import { FormattedMessage, useIntl } from 'react-intl'
 import { EmptyState } from 'vtex.styleguide'
 
-defineMessages({
-  emptyStateTitle: {
-    id: 'docs/empty.title',
-    defaultMessage: '',
-  },
-})
+const EmptyDocs: FC = () => {
+  const intl = useIntl()
 
-const EmptyDocs: FC<InjectedIntlProps> = ({ intl }) => (
-  <div className="pa7 w-100 flex justify-center">
-    <EmptyState title={intl.formatMessage({ id: 'docs/empty.title' })}>
-      <p>
-        <FormattedMessage id="docs/empty-docs" />
-      </p>
-    </EmptyState>
-  </div>
-)
+  return (
+    <div className="pa7 w-100 flex justify-center">
+      <EmptyState title={intl.formatMessage({ id: 'docs/empty.title' })}>
+        <p>
+          <FormattedMessage id="docs/empty-docs" />
+        </p>
+      </EmptyState>
+    </div>
+  )
+}
 
-export default injectIntl(EmptyDocs)
+export default EmptyDocs